archived unused files and clean code
