+++ .gitignore
@@ -2,6 +2,8 @@
 __pycache__/
 .idea/
 .idea/workspace.xml
+*.jsonl
+visited_urls.txt
 cert.pem
 key.pem
 offload/
