+++ pinecone_upload.py
