+++ utils/custom_logger.py
@@ -2,7 +2,7 @@ import os
 import sys
 import datetime
 from logging import Logger, StreamHandler, Formatter, FileHandler
-from typing import Self
+from typing_extensions import Self
 from pathlib import Path
